Move name generation to StrawberryConfig
